Improved code quality, Added documentation strings
