--- amcp_pylib/response/client_error_response.py
@@ -1,5 +1,0 @@
-from .base import Base as ResponseBase
-
-
-class ClientErrorResponse(ResponseBase):
-    pass
