--- amcp_pylib/response/info_response.py
@@ -1,5 +1,0 @@
-from .base import Base as ResponseBase
-
-
-class InfoResponse(ResponseBase):
-    pass
