--- amcp_pylib/response/server_error_response.py
@@ -1,5 +1,0 @@
-from .base import Base as ResponseBase
-
-
-class ServerErrorResponse(ResponseBase):
-    pass
