--- amcp_pylib/response/success_response.py
@@ -1,5 +1,0 @@
-from .base import Base as ResponseBase
-
-
-class SuccessResponse(ResponseBase):
-    pass
