+++ amcp_pylib/response/types/client_error_response.py
@@ -1,0 +1,5 @@
+from ..response_base import ResponseBase as ResponseBase
+
+
+class ClientErrorResponse(ResponseBase):
+    pass
