+++ amcp_pylib/response/types/info_response.py
@@ -1,0 +1,5 @@
+from ..response_base import ResponseBase as ResponseBase
+
+
+class InfoResponse(ResponseBase):
+    pass
