+++ amcp_pylib/response/types/server_error_response.py
@@ -1,0 +1,5 @@
+from ..response_base import ResponseBase as ResponseBase
+
+
+class ServerErrorResponse(ResponseBase):
+    pass
