+++ amcp_pylib/response/types/success_response.py
@@ -1,0 +1,5 @@
+from ..response_base import ResponseBase as ResponseBase
+
+
+class SuccessResponse(ResponseBase):
+    pass
